+++ src/TypeCasters/EnumerableKeyCaster.php
@@ -4,8 +4,9 @@ namespace Somnambulist\Components\AttributeModel\TypeCasters;
 
 use InvalidArgumentException;
 use Somnambulist\Components\AttributeModel\Contracts\AttributeCasterInterface;
-use Somnambulist\Components\Domain\Entities\AbstractEnumeration;
-use Somnambulist\Components\Domain\Entities\AbstractMultiton;
+use Somnambulist\Components\Models\AbstractEnumeration;
+use Somnambulist\Components\Models\AbstractMultiton;
+
 use function in_array;
 use function is_a;