+++ src/TypeCasters/ExternalIdentityCaster.php
@@ -3,7 +3,7 @@
 namespace Somnambulist\Components\AttributeModel\TypeCasters;
 
 use Somnambulist\Components\AttributeModel\Contracts\AttributeCasterInterface;
-use Somnambulist\Components\Domain\Entities\Types\Identity\ExternalIdentity;
+use Somnambulist\Components\Models\Types\Identity\ExternalIdentity;
 
 final class ExternalIdentityCaster implements AttributeCasterInterface
 {