+++ tests/ModelTest.php
@@ -5,7 +5,8 @@ namespace Somnambulist\Components\AttributeModel\Tests;
 use BadMethodCallException;
 use PHPUnit\Framework\TestCase;
 use Somnambulist\Components\AttributeModel\Tests\Support\Stubs\Models\User;
-use Somnambulist\Components\Domain\Entities\Types\DateTime\DateTime;
+use Somnambulist\Components\Models\Types\DateTime\DateTime;
+
 use function date;
 use function password_hash;